python-ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.0
sentence-transformers==2.3.1
huggingface-hub>=0.20.0
chromadb==0.4.22
pydantic==2.6.0
aiohttp==3.9.1
pymongo==4.6.0
redis==5.0.1
python-multipart==0.0.6
torch>=2.0.0


